# Conflicts:
#	C/c4Query.hh
#	C/tests/c4DatabaseInternalTest.cc
#	C/tests/c4DatabaseTest.cc
#	C/tests/c4DocumentTest.cc
#	C/tests/c4Test.hh
#	LiteCore/Support/LogDecoder.cc
#	LiteCore/Support/Logging.hh
#	LiteCore/Support/ThreadedMailbox.hh
#	LiteCore/tests/LiteCoreTest.cc
#	Replicator/Pusher+Revs.cc
#	Replicator/tests/ReplicatorLoopbackTest.cc
#	Replicator/tests/ReplicatorLoopbackTest.hh
#	Replicator/tests/ReplicatorSGTest.cc
#	vendor/fleece

+++ LiteCore/Support/LogEncoder.cc
@@ -214,7 +214,7 @@ namespace litecore {
                         if (sizeof(param) == 8)
                             param = fleece::endian::encLittle64(param);
                         else
-                            param = fleece::endian::encLittle32(param);
+                            param = fleece::endian::encLittle32((uint32_t)param);
                         _writer.write(&param, sizeof(param));
                         break;
                     }
